Update signatures for consistency
This is a horrible set of changes, but now all the signatures are
generated on the same machine, with 64-bit integers for the fields
that vary between environments. This makes them consistent with
each other and with the attribute casting in the signature tests.
Also updating some signatures for the recent changes related to
pynwb 1.2.1 (e.g. column names in epochs, not storing spec).