Add a full test for ROIs in the new LabView format
This comes from subsampling one of the test experiments. The test
input and signature don't include any imaging data just yet, to
keep the file size small.
The point of this test is to check that we reference ROIs correctly
even when those ROIs on the same z-plane are not consecutive in
the ROI.dat table. The referencing happens from the main code
(nwb_file.py), so we need a test for the whole import process.