Force 32 bits for grid spacing for consistency
Some environments would write this as 64 bits, making signature
comparisons harder (and this could not be fixed by casting the
value before the signature, as the value itself would have changed,
not just the number of bits used).
This changes the spacing of the grid slightly compared to when we
were storing the manifold explicitly, but I think the differences
are at atomic scale...
Also updating the signatures for the larger files.