Allow creating ROI readers for newest LabView
To have full access to the header information (not used yet, but
it will be to check for variable resolution/shape), we need to
create the reader earlier on, since we don't store the header
object anywhere. This also lets us use the reader across
different methods if required (e.g when adding the acquisitions
for each ROI).