adapt to possible lack of speed data file
this can (at least currently) only happen in post-2018 versions (2.3.1
and 3.0.0) of LabView, because the latter have trial information in the
header files. In the pre-2018, the speed data is the only way a
deduction of trial times is possible, so in this case we can't do
without a speed data file (and a ValueError is raised.)

This set of changes involves transferring some calculations from the
epochs (which contain the speed data) to the trials in the file. A
side-effect of this are small signature changes, due to the 1e-9 small
difference in start and stop times between epoch and trial.