Put the TDMS data in the right order
May be wrong!

Previous version had all data from the first ROI
followed by the second etc, but the original TDMS stores data
in complete cycles (first cycle for all ROIs, then second cycle
for all ROIs etc).

This is a bit less general than the previous version,
which could have been easily adapted for non-consecutive ROIs.
However, the whole file is based on the assumption that we are
keeping only the first N ROIs, so the simplification is probably
worth the loss of generality in one function...